requirements.txt
langchain-core
langchain-openai
langchain-community
openai
python-dotenv
pydantic
fastapi
uvicorn
streamlit
pymongo
langchain>=0.1.0
langchain-pinecone>=0.1.0
pinecone-client>=3.0.0
python-multipart>=0.0.6
pypdf>=3.17.1
tiktoken>=0.5.1
pillow>=10.0.0
pytesseract>=0.3.10